requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cycler==0.11.0
fonttools==4.39.3
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==23.0
pandas==2.0.0
pandas-stubs==1.2.0.61
Pillow==9.5.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
